+++ armsrc/iso14443a.h
@@ -168,13 +168,15 @@ bool SimulateIso14443aInit(uint8_t tagType, uint16_t flags, uint8_t *data,
                            uint8_t *ulc_key);
 
 bool GetIso14443aCommandFromReader(uint8_t *received, uint16_t received_maxlen, uint8_t *par, int *len);
-void iso14443a_antifuzz(uint32_t flags);
+void iso14443a_antifuzz(uint32_t flags, uint8_t mode);
 int GetATQA(uint8_t *resp, uint16_t resp_len, uint8_t *resp_par, const iso14a_polling_parameters_t *polling_parameters);
 void ReaderIso14443a(PacketCommandNG *c);
 void ReaderTransmit(const uint8_t *frame, uint16_t len, uint32_t *timing);
 void ReaderTransmitBitsPar(const uint8_t *frame, uint16_t bits, uint8_t *par, uint32_t *timing);
 void ReaderTransmitPar(const uint8_t *frame, uint16_t len, uint8_t *par, uint32_t *timing);
 uint16_t ReaderReceive(uint8_t *receivedAnswer, uint16_t answer_maxlen, uint8_t *par);
+uint32_t iso14a_last_tx_end(void);
+uint32_t iso14a_last_rx_start(void);
 
 void iso14443a_setup(uint8_t fpga_minor_mode);
 int iso14_apdu(uint8_t *cmd, uint16_t cmd_len, bool send_chaining, void *data, uint16_t data_len, uint8_t *res);